+++ Sources/Utils/CHANGELOG.md
@@ -1,3 +1,6 @@
+### 0.18.0 (pre-release):
+* [Change] Improve support for compound persistent fields.
+* [Enhancement] Support (de)serialization of `IConfigNode` types in [`PersistentFieldAttribute`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_ConfigUtils_PersistentFieldsFileAttribute.htm).
 ### 0.17.0 (Nov 19, 2016):
 * [Fix] Improved code samples and fixed some docs.
 * [Enhancement] `MessageBoolValue` class to format boolean values.